Mock everything by default in UnitTestCase::mock()
Instead of passing in the methods you want to mock, now you pass in the
methods you don't want to mock (the opposite of previous behaviour).
This makes it easier to ensure you don't accidentally pass through to
the implemented behaviour; rather, any "forgotten" methods become
mocked with no expectations, and so will raise a Mockery exception.