Fix inheritance of base models
Using "$ref" in an parameter definition doesn't allow anything to be
overridden to be different to the model definition. So the models that
are defined need to be exactly what gets used as the parameter -- no
changes can be made in the parameter definition (i.e. it doesn't work
the same as the "extends" functionality).