Add "_query" named match to link patterns
This will extract query string parameters into the main parameters, as
returned from ResourceBuilder::parseHref().

By adding a named capture group of "_query" to a link pattern, which
captures the entire query string, all of the query string parameters
can be extracted as parameters to be used on the resulting command.