fix issue with validating commands received from next links
Issue would arrise when using the ResourceIterator class to
iterate over a set of searched cases. The since_created_at,
max_created_at, since_updated_at, and max_updated_at were setup
to expect a DateTime instance. However in the case a link was
used it would already contain the filtered value. This would throw
an exception since the value was not an instance of DateTime.

Resolved by removing the requirement to be a DateTime instance.
Instead the filter Date::toTimestamp takes care of the different
possible values that could arrise.